[FLINK-XXXX] Move mutli-threaded `streamFlatMap` to new class
In order to organize and separate the logic of the single and multi-threaded
`StreamFlatMap` class, it is cleaner to seprate each in a different class
as the logic of the multi-threaded `StreamFlatMap` started to devaite from
the logic of the existing single threaded `StreamFlatMap`.
